Bugfix to the overlap check, during initial particle placement.
This is critical for the function of Metropolis-Hastings, because if
particles ever happen to overlap, the quantum force will be so large
that suggested moves are practically out of bounds and system gets stuck
rejecting every attempt to leave the overlapping state.

Normally during simulation patricles will not transfer into these
forbidden states, so this seems to have been an issue only with the
initialiaztion.